+++ source/mibrr/src/MibrrData.hpp
@@ -1,12 +1,12 @@
 // Title:    Header file for MibrrData Class
 // Author:   Kyle M. Lang
 // Created:  2014-AUG-24
-// Modified: 2017-SEP-30
+// Modified: 2017-OCT-27
 // Purpose:  This class contains data- and sampling-related functions used by the
 //           MIBRR Gibbs sampler.
 
 //--------------------- COPYRIGHT & LICENSING INFORMATION ---------------------//
-//  Copyright (C) 2017 Kyle M. Lang <kyle.lang@ttu.edu>                        //  
+//  Copyright (C) 2017 Kyle M. Lang <k.m.lang@uvt.nl>                          //  
 //                                                                             //
 //  This file is part of mibrr.                                                //
 //                                                                             //
@@ -74,7 +74,11 @@ public:
   double getDataScales(int) const;
   // @param:  a column index
   // @return: the  scale of the data in the specified column
-  
+
+  vector<int> getMissIndices(int targetIndex) const;
+  // @param:  a column index
+  // @return: the row indices of missing values in the specified column
+
   MatrixXd getData() const;
   // @return: the data matrix